# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  version: v1.0.0
  title: Attendance Application API
  description: API for Attendance application.
  x-audience: external-partner
servers:
  - url: 'http://{host}:{port}'
    description: http server
    variables:
      host:
        default: 'localhost'
        description: identity server host
      port:
        default: '8000'
        description: identity server port
  - url: 'https://{host}'
    description: https server
    variables:
      host:
        default: 'localhost'
        description: identity server domain
security:
  - AccessToken: []
...
paths:
...
  /api/v1/users:
    get:
      summary: Get list of users
      operationId: getAllUsers
      deprecated: false
      tags:
        - Users
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Sort"
        - $ref: "#/components/parameters/UserSearch"
        - $ref: "#/components/parameters/UserIdSearch"
      responses:
        "200":
          description: list of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetUsersResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/UnexpectedError"
        "403":
          $ref: "#/components/responses/Forbidden"
        "401":
          $ref: "#/components/responses/UnAuthorized"
    post:
      summary: Create a new user
      operationId: createUser
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/UnexpectedError"
        "403":
          $ref: "#/components/responses/Forbidden"
        "401":
          $ref: "#/components/responses/UnAuthorized"
  /api/v1/users/me:
    get:
      summary: Get logged in user details
      operationId: getLoggedInUserDetails
      tags:
        - Users
      responses:
        "200":
          description: fetch user details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfileResponse"
        "500":
          $ref: "#/components/responses/UnexpectedError"
        "403":
          $ref: "#/components/responses/Forbidden"
        "401":
          $ref: "#/components/responses/UnAuthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/users/{id}:
    get:
      summary: Get user details using id
      operationId: getUserById
      parameters:
        - name: "id"
          in: "path"
          description: "The user's id that needs to be fetched"
          required: true
          schema:
            type: "integer"
            format: "int64"
      tags:
        - Users
      responses:
        "200":
          description: fetch user details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDetailedResponse"
        "500":
          $ref: "#/components/responses/UnexpectedError"
        "403":
          $ref: "#/components/responses/Forbidden"
        "401":
          $ref: "#/components/responses/UnAuthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      summary: Update user by id
      operationId: updateUser
      parameters:
        - name: "id"
          in: "path"
          description: "The user's id that needs to be updated"
          required: true
          schema:
            type: "integer"
            format: "int64"
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        "200":
          description: User details updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/UnexpectedError"
        "404":
          $ref: "#/components/responses/NotFound"
        "403":
          $ref: "#/components/responses/Forbidden"
        "401":
          $ref: "#/components/responses/UnAuthorized"
    delete:
      summary: Soft delete users
      operationId: softDeleteUser
      parameters:
        - name: "id"
          in: "path"
          description: "The user's id that needs to be soft deleted"
          required: true
          schema:
            type: "integer"
            format: "int64"
      tags:
        - Users
      responses:
        "200":
          description: User soft deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/UnexpectedError"
        "404":
          $ref: "#/components/responses/NotFound"
        "403":
          $ref: "#/components/responses/Forbidden"
        "401":
          $ref: "#/components/responses/UnAuthorized"
  /api/v1/internal-users/{id}:
    get:
      summary: Get user details using id
      operationId: getInternalUserById
      parameters:
        - name: "id"
          in: "path"
          description: "The user's id that needs to be fetched"
          required: true
          schema:
            type: "integer"
            format: "int64"
      security:
        - BasicAuth: [ ]
      tags:
        - Users
      responses:
        "200":
          description: fetch user details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDetailedResponse"
        "500":
          $ref: "#/components/responses/UnexpectedError"
        "403":
          $ref: "#/components/responses/Forbidden"
        "401":
          $ref: "#/components/responses/UnAuthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/users/{user-id}/assign-roles:
    post:
      summary: Assign roles to user
      operationId: assignRolesToUser
      parameters:
        - name: "user-id"
          in: "path"
          description: "The user's id to whom the roles to be assigned"
          required: true
          schema:
            type: "integer"
            format: "int64"
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignRolesToUserRequest'
      responses:
        "200":
          description: Successful user role updation response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/UnexpectedError"
        "404":
          $ref: "#/components/responses/NotFound"
        "403":
          $ref: "#/components/responses/Forbidden"
        "401":
          $ref: "#/components/responses/UnAuthorized"
...
components:
  securitySchemes:
    AccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: The authorization token used by the user to authorize API requests.
    RefreshToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: The token used to refresh an expired access token.
    BasicAuth:
      type: http
      scheme: basic
      description: Basic authentication using username and password.
  parameters:
    Page:
      in: query
      name: page
      schema:
        type: integer
        format: int32
        default: 0
      required: false
      description: the page number.
    PageSize:
      in: query
      name: size
      schema:
        type: integer
        format: int32
        default: 20
      required: false
      description: the size of a page
    Sort:
      in: query
      name: sort
      schema:
        type: string
        default: id,asc
      required: false
      example: id,asc
      description: sort
...
    UserSearch:
      in: query
      name: search
      schema:
        type: string
      required: false
      description: This adds a string search to the following fields.
        - name
        - username
        - email
        - designation
        - epn
...
    UserIdSearch:
      in: query
      name: ids
      schema:
        type: array
        items:
          type: "integer"
          format: "int64"
      required: false
      description: This adds a search to the id field.
        - ids
...
  schemas:
...
    ApiError:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: The server time at which this response was created.
        code:
          type: integer
          format: int32
          description: The status code the server returned.
        status:
          type: boolean
          description: Whether or not the request was a success or not.
          default: false
        message:
          type: string
          description: A short error message from server. If there are multiple errors they will be comma seperated
    UserResponse:
      type: "object"
      properties:
        timestamp:
          type: "string"
          format: "date-time"
          description: The server time at which this response was created.
        code:
          type: "integer"
          format: "int32"
          description: The status code the server returned.
        message:
          type: "string"
          description: A short error message from server. If there are multiple errors they will be comma seperated
        status:
          type: "boolean"
          description: Whether or not the request was a success or not.
        data:
          type: "array"
          items:
            $ref: "#/components/schemas/UserDto"
    UserProfileResponse:
      type: "object"
      properties:
        timestamp:
          type: "string"
          format: "date-time"
          description: The server time at which this response was created.
        code:
          type: "integer"
          format: "int32"
          description: The status code the server returned.
        message:
          type: "string"
          description: A short error message from server. If there are multiple errors they will be comma seperated
        status:
          type: "boolean"
          description: Whether or not the request was a success or not.
        data:
          type: "array"
          items:
            $ref: "#/components/schemas/UserProfileDto"
    UserDto:
      type: object
      properties:
        id:
          type: "integer"
          format: "int64"
        name:
          type: "string"
        email:
          type: "string"
        designation:
          type: "string"
        userName:
          type: "string"
        epn:
          type: "string"
        enabled:
          type: "boolean"
        mobileNumber:
          type: "string"
          example: "000000000"
        deleted:
          type: boolean
        userType:
          $ref: "#/components/schemas/UserType"
    UserProfileDto:
      type: object
      properties:
        id:
          type: "integer"
          format: "int64"
        name:
          type: "string"
        email:
          type: "string"
        designation:
          type: "string"
        userName:
          type: "string"
        epn:
          type: "string"
        enabled:
          type: "boolean"
        deleted:
          type: boolean
        userType:
          $ref: "#/components/schemas/UserType"
        mobileNumber:
          type: "string"
        Privileges:
          type: array
          items:
            type: string
    GetUsersResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: The server time at which this response was created.
        code:
          type: integer
          format: int32
          description: The status code the server returned.
          example: 200
        status:
          type: boolean
          description: Whether or not the request was a success or not.
          default: true
        message:
          type: string
          description: A short success message from server
        data:
          type: array
          items:
            $ref: "#/components/schemas/UserPage"
    UserPage:
      type: object
      properties:
        page:
          type: integer
          format: int64
          description: The current page number.
        size:
          type: integer
          format: int32
          description: The size of current page.
        first:
          type: boolean
          description: whether current page is the first one.
        last:
          type: boolean
          description: whether current page is the last one.
        totalPages:
          type: integer
          format: int64
          description: total no:of pages available
        totalElements:
          type: integer
          format: int64
          description: total no:of elements available.
        data:
          type: array
          items:
            $ref: "#/components/schemas/UserDto"
...
    PrivilegeDetailedDto:
      type: object
      properties:
        id:
          type: "integer"
          format: "int64"
        name:
          type: "string"
        description:
          type: "string"
        application:
          type: array
          items:
            $ref: "#/components/schemas/ApplicationDto"
            type: ApplicationDto
...
    ApplicationDto:
      type: object
      properties:
        id:
          type: "integer"
          format: "int64"
        name:
          type: "string"
        deleted:
          type: boolean
...
    RoleDetailedDto:
      type: object
      properties:
        id:
          type: "integer"
          format: "int64"
        name:
          type: "string"
        description:
          type: "string"
        deleted:
          type: "boolean"
        privilege:
          type: array
          items:
            $ref: "#/components/schemas/PrivilegeDetailedDto"
...
    UserDetailedResponse:
      type: "object"
      properties:
        timestamp:
          type: "string"
          format: "date-time"
          description: The server time at which this response was created.
        code:
          type: "integer"
          format: "int32"
          description: The status code the server returned.
        message:
          type: "string"
          description: A short error message from server. If there are multiple errors they will be comma seperated
        status:
          type: "boolean"
          description: Whether or not the request was a success or not.
        data:
          type: "array"
          items:
            $ref: "#/components/schemas/UserDetailedDto"
    UserDetailedDto:
      type: object
      properties:
        id:
          type: "integer"
          format: "int64"
        name:
          type: "string"
        email:
          type: "string"
        designation:
          type: "string"
        userName:
          type: "string"
        epn:
          type: "string"
        mobileNumber:
          type: "string"
          example: "000000000"
        enabled:
          type: "boolean"
        deleted:
          type: "boolean"
        roles:
          type: array
          items:
            $ref: "#/components/schemas/RoleDetailedDto"
...
    CreateUserRequest:
      type: object
      required:
        - name
        - designation
        - email
        - epn
      properties:
        name:
          type: string
          description: "user's name"
          x-required-message: "{messages.nameIsRequired}"
        designation:
          type: string
          description: designation of user
          x-required-message: "{messages.designationIsRequired}"
        email:
          type: string
          format: email
          description: "user email"
          x-required-message: "{messages.emailIsRequired}"
          x-email-message: "{messages.emailShouldBeValid}"
        epn:
          type: string
          pattern: "^[0-9]+$"
          description: "user EPN number"
          x-required-message: "{messages.epnIsRequired}"
          x-pattern-message: "{messages.epnShouldBeNumbers}"
        mobileNumber:
          type: "string"
          example: "000000000"
          description: "mobile number of user"
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
          description: "user's name"
        designation:
          type: string
          description: "user's designation"
        mobileNumber:
          type: "string"
          example: "000000000"
          description: "mobile number of user"
...
    AssignRolesToUserRequest:
      type: object
      required:
        - roleIds
      properties:
        roleIds:
          type: array
          description: Roles to be assigned to users
          items:
            type: integer
            format: int64
          x-required-message: "{messages.roleIdsAreRequired}"
...
    GenericResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: The server time at which this response was created.
        code:
          type: integer
          format: int32
          description: The status code the server returned.
          example: 200
        status:
          type: boolean
          description: Whether or not the request was a success or not.
          default: true
        message:
          type: string
          description: A short success message from server
...
    UserType:
      type: string
      enum:
        - CLIENT
        - INTERNAL-USER
  responses:
    BadRequest:
      description: Request syntax is invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            timestamp: 2023-02-21T18:53:37.709Z
            code: 400
            status: false
            message: "bad request"
    UnAuthorized:
      description: Authorization information is missing or invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            timestamp: 2023-02-21T18:53:37.709Z
            code: 401
            status: false
            message: "unauthorized"
    NotFound:
      description: Requested resource was not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            timestamp: 2023-02-21T18:53:37.709Z
            code: 404
            status: false
            message: "not found"
    UnexpectedError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            timestamp: 2023-02-21T18:53:37.709Z
            code: 500
            status: false
            message: "something unexpected occured"
    Forbidden:
      description: Forbidden error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            timestamp: 2023-02-21T18:53:37.709Z
            code: 403
            status: false
            message: "forbidden"